+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.1.0
 
 * Added support for raw `String` literals.
+* Automatically escapes single quotes in now-raw `String` literals.
 
 ## 2.0.0